Menu fade in / out transition (#14)
Fade in the main mav menu using a CSS transition.

This technique works, although the first transition causes the fonts to 'flash' when run in dev mode. Apparently this is a styled components global includes issue. The current workaround is to import your fonts outside of styled components. Given this only affects development, this can be overlooked for now.